src/requirements.txt
pandas
requests
PyYAML
DateTime
regex